app_rag/requirements.txt
databricks-sdk
openai
streamlit
